chore: move wallet connect context (#857)
