+++ packages/renderer/src/renderer/contexts/import/WalletConnect/defaults.ts
@@ -2,18 +2,18 @@
 // SPDX-License-Identifier: GPL-3.0-only
 /* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
 
-import type { WalletConnectContextInterface } from './types';
+import type { WalletConnectImportContextInterface } from './types';
 
-export const defaultWalletConnectContext: WalletConnectContextInterface = {
-  wcConnecting: false,
-  wcDisconnecting: false,
-  wcFetchedAddresses: [],
-  wcInitialized: false,
-  wcNetworks: [],
-  wcSessionRestored: false,
-  connectWc: () => new Promise(() => {}),
-  disconnectWcSession: () => new Promise(() => {}),
-  fetchAddressesFromExistingSession: () => {},
-  setWcFetchedAddresses: () => {},
-  setWcNetworks: () => {},
-};
+export const defaultWalletConnectImportContext: WalletConnectImportContextInterface =
+  {
+    isImporting: false,
+    wcFetchedAddresses: [],
+    wcNetworks: [],
+    getSelectedAddresses: () => [],
+    handleConnect: () => new Promise(() => {}),
+    handleDisconnect: () => new Promise(() => {}),
+    handleFetch: () => {},
+    setWcFetchedAddresses: () => {},
+    handleImportProcess: () => new Promise(() => {}),
+    setWcNetworks: () => {},
+  };
