+++ packages/renderer/src/renderer/contexts/main/WalletConnect/defaults.ts
@@ -1,0 +1,12 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
+
+import type { WalletConnectContextInterface } from './types';
+
+export const defaultWalletConnectContext: WalletConnectContextInterface = {
+  wcSessionRestored: false,
+  connectWc: () => new Promise(() => {}),
+  disconnectWcSession: () => new Promise(() => {}),
+  fetchAddressesFromExistingSession: () => {},
+};
