+++ packages/renderer/src/renderer/screens/Import/WalletConnect/Import/index.tsx
@@ -23,11 +23,9 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 
 /** Temp */
-import { useAccountStatuses } from '@app/contexts/import/AccountStatuses';
 import { useAddresses } from '@app/contexts/import/Addresses';
 import { useConnections } from '@app/contexts/common/Connections';
-import { useImportHandler } from '@app/contexts/import/ImportHandler';
-import { useWalletConnect } from '@app/contexts/import/WalletConnect';
+import { useWalletConnectImport } from '@app/contexts/import/WalletConnect';
 import { useEffect, useState } from 'react';
 import { ellipsisFn } from '@w3ux/utils';
 import { WcFlexRow, WcSessionButton } from './Wrappers';
@@ -39,40 +37,39 @@ import {
 import type { ImportProps } from './types';
 
 export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
-  const { insertAccountStatus } = useAccountStatuses();
+  const {
+    darkMode,
+    getOnlineMode,
+    wcSyncFlags: {
+      wcConnecting,
+      wcDisconnecting,
+      wcInitialized,
+      wcSessionRestored,
+    },
+  } = useConnections();
+
   const { isAlreadyImported, wcAddresses } = useAddresses();
-  const { darkMode, getOnlineMode } = useConnections();
-  const { handleImportAddress } = useImportHandler();
 
   const theme = darkMode ? themeVariables.darkTheme : themeVariables.lightThene;
   const {
-    connectWc,
-    disconnectWcSession,
-    fetchAddressesFromExistingSession,
-    setWcFetchedAddresses,
-    setWcNetworks,
-    wcConnecting,
-    wcDisconnecting,
+    isImporting,
     wcFetchedAddresses,
-    wcInitialized,
     wcNetworks,
-    wcSessionRestored,
-  } = useWalletConnect();
+    getSelectedAddresses,
+    handleConnect,
+    handleDisconnect,
+    handleFetch,
+    setWcFetchedAddresses,
+    handleImportProcess,
+    setWcNetworks,
+  } = useWalletConnectImport();
 
-  const [isImporting, setIsImporting] = useState(false);
   const [accordionActiveIndices, setAccordionActiveIndices] = useState<
     number[]
   >(Array.from({ length: 2 }, (_, index) => index));
 
   const getSelectedNetworkCount = () =>
     wcNetworks.filter(({ selected }) => selected).length;
-
-  /**
-   * Handle connect button click.
-   */
-  const handleConnect = async () => {
-    await connectWc();
-  };
 
   /**
    * Handle address checkbox click.
@@ -97,12 +94,6 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
   };
 
   /**
-   * Get the selected addresses to import.
-   */
-  const getSelectedAddresses = () =>
-    wcFetchedAddresses.filter(({ selected }) => selected);
-
-  /**
    * Get import button label text.
    */
   const getImportLabel = () => {
@@ -124,37 +115,6 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
       <span>Currently offline. Please go online to enable connections.</span>
     </InfoCard>
   );
-
-  /**
-   * Handle importing the selected WalletConnect addresses.
-   */
-  const handleImportProcess = async () => {
-    const selectedAddresses = getSelectedAddresses();
-    if (selectedAddresses.length === 0) {
-      return;
-    }
-
-    setIsImporting(true);
-    for (const selected of selectedAddresses) {
-      const { encoded } = selected;
-
-      if (isAlreadyImported(encoded)) {
-        continue;
-      }
-
-      const accountName = ellipsisFn(encoded);
-      await handleImportAddress(encoded, 'wallet-connect', accountName, false);
-      insertAccountStatus(encoded, 'wallet-connect');
-    }
-
-    setIsImporting(false);
-    setShowImportUi(false);
-
-    /** Clear selected WalletAccount addresses. */
-    setWcFetchedAddresses((prev) =>
-      prev.map((item) => ({ ...item, selected: false }))
-    );
-  };
 
   /**
    * Effects.
@@ -239,7 +199,7 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
                           wcConnecting ||
                           wcDisconnecting
                         }
-                        onClick={async () => await disconnectWcSession()}
+                        onClick={async () => await handleDisconnect()}
                       >
                         Disconnect
                       </WcSessionButton>
@@ -252,7 +212,7 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
                           wcDisconnecting
                         }
                         onClick={() => {
-                          fetchAddressesFromExistingSession();
+                          handleFetch();
                           setAccordionActiveIndices([1]);
                         }}
                       >
@@ -393,7 +353,9 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
                         disabled={
                           isImporting || getSelectedAddresses().length === 0
                         }
-                        onClick={async () => await handleImportProcess()}
+                        onClick={async () =>
+                          await handleImportProcess(setShowImportUi)
+                        }
                       >
                         {getImportLabel()}
                       </button>
